requirements.txt
scipy==1.2.0
tqdm>=4.65.0
torchvision>=0.14.1
pillow>=9.4.0
imageio>=2.28.1
timm>=0.6.13
einops>=0.6.1
natsort>=8.3.1